Fixing tests for live pricing/routing model
