+++ tests/ganache/test_frozen_position_and_blacklist.py
@@ -239,7 +239,7 @@ def hot_wallet(web3: Web3, busd_token: Contract, hot_wallet_private_key: HexByte
 @pytest.fixture()
 def strategy_path() -> Path:
     """Where do we load our strategy file."""
-    return Path(os.path.join(os.path.dirname(__file__), "strategies", "frozen_asset.py"))
+    return Path(os.path.join(os.path.dirname(__file__), "../strategies", "frozen_asset.py"))
 
 
 @pytest.fixture()
@@ -292,7 +292,7 @@ def universe_model(universe, supported_reserves) -> StaticUniverseModel:
 @pytest.fixture()
 def strategy_path() -> Path:
     """Where do we load our strategy file."""
-    return Path(os.path.join(os.path.dirname(__file__), "strategies", "frozen_asset.py"))
+    return Path(os.path.join(os.path.dirname(__file__), "../strategies", "frozen_asset.py"))
 
 
 @pytest.fixture()
