+++ tradeexecutor/strategy/qstrader/portfolio_construction_model.py
@@ -291,7 +291,8 @@ class PortfolioConstructionModel:
 
         # Get prices for existing assets so we have some idea how much they sell for
         for asset_id, asset_data in current_portfolio.items():
-            target_prices[asset_id] = self.pricing_model.get_buy_price(dt, asset_id)
+            pair = self.pricing_model.get_pair_for_id(asset_id)
+            target_prices[asset_id] = self.pricing_model.get_buy_price(dt, pair, None)
 
         # Expose internal states to unit tests
         debug_details["positions_at_start_of_construction"] = current_portfolio.copy()  # current_portfolio is mutated later
